fix(urls_table): Use variable flag to set whether `index.html` file
This patch helps to reduce call to `fs_path_to_url` when mapping url
from given path. We can simply use variable to set whether we already
found `index.html` file or not. Which will be use to decide whether
directory listing page need to build for given path or not.

Signed-off-by: PrajwalCH <prajwal.chapagain58@gmail.com>